Add tags

+++ common/src/main/java/com/tencent/rss/common/config/ConfigUtils.java
@@ -178,12 +178,12 @@ public class ConfigUtils {
     return configOptionList;
   }
 
-  public static Function<Long, Boolean> positiveLongValidator = value -> value > 0L;
+  public static Function<Long, Boolean> positiveLongValidator = value -> value > 0;
 
-  public static Function<Long, Boolean> non_negativeLongValidator = value -> value >= 0L;
+  public static Function<Long, Boolean> non_negativeLongValidator = value -> value >= 0;
 
-  public static Function<Integer, Boolean> positiveIntegerValidator =
-      value -> value > 0 && value <= Integer.MAX_VALUE;
+  public static Function<Long, Boolean> positiveIntegerValidator =
+      value -> value > 0L && value <= Integer.MAX_VALUE;
 
   public static Function<Integer, Boolean> positiveIntegerValidator2 =
       value -> value > 0;
